requirements.txt
streamlit
pandas==2.1.3
numpy
matplotlib
openpyxl==3.1.2
json5==0.9.14
altair==5.1.2
datetime
st-pages==0.4.5
python-dateutil==2.8.2
st-pages==0.4.5